Lab 14 task 4 fixed

+++ task4-sender.c
@@ -4,6 +4,12 @@
 #include <unistd.h>
 #include <stdlib.h>
 
+int cant_send;
+
+void allow_send(int nsig) {
+  cant_send = 0;
+}
+
 int main(void) {
   pid_t my_pid, receiver_pid;
   int   input, mask;
@@ -13,20 +19,30 @@ int main(void) {
   printf("My PID = %d\n", my_pid);
 
   printf("Enter receiver PID: ");
-  scanf("%d", &receiver_pid);
+  if (scanf("%d", &receiver_pid) < 0) {
+    printf("Invalid PID\n");
+    exit(-1);
+  }
 
   printf("Enter an integer value to send: ");
-  scanf("%d", &input);
+  if (scanf("%d", &input) < 0) {
+    printf("Invalid integer value\n");
+    exit(-1);
+  }
+
+  (void) signal(SIGUSR1, allow_send);
 
   mask = 1;
   while (mask != 0) {
+    while(cant_send); // Wait for client to respond.
+    cant_send = 1;
+
     if (input & mask) {
       kill(receiver_pid, SIGUSR2);
     } else {
       kill(receiver_pid, SIGUSR1);
     }
     mask <<= 1;
-    for (i = 0; i < 30000000; ++i);
   }
 
   return 0;
